Adjust code to be able to boot up for Tulsa County

+++ lib/tasks/save_case.rake
@@ -7,9 +7,8 @@ namespace :save do
     court_cases.each do |c|
       data = Nokogiri::HTML.parse(c.case_html.html)
       errors = data.css('.errorMessage')
-      if errors.count > 0
-        c.case_html.destroy
-      end
+
+      c.case_html.destroy if errors.count.positive?
       bar.increment!
     end
   end
@@ -23,18 +22,9 @@ namespace :save do
       parsed_html = Nokogiri::HTML(c.case_html.html.body)
       data = OscnScraper::Parsers::BaseParser.new(parsed_html).build_object
       logs = Importers::Logger.new(c)
-      judge_id = Matchers::Judge.new(data[:judge]).judge_id
+      judge_id = Matchers::Judge.new(c.county_id, data[:judge]).judge_id
       c.update(current_judge_id: judge_id)
       bar.increment!
-    end
-  end
-
-  task update_year: [:environment] do
-    cases = CourtCase.where(filed_on: Date.new(2020,7,1)..Date.new(2021,7,1)).where("case_number LIKE 'C%'")
-    bar = ProgressBar.new(cases.count)
-    cases.each do |c|
-      bar.increment!
-      CourtCaseWorker.perform_async(c.case_number)
     end
   end
 end
